numeric ristriction

+++ app/src/main/java/com/chetz/myapplication/Math.java
@@ -2,23 +2,23 @@ package com.chetz.myapplication;
 
 public class Math {
 
-    public int addition(int one, int two) {
-        int result = one + two;
+    public double addition(double one, double two) {
+        double result = one + two;
         return result;
     }
 
-    public int subtraction(int one, int two) {
-        int result = one - two;
+    public double subtraction(double one, double two) {
+        double result = one - two;
         return result;
     }
 
-    public int multiplication(int one, int two) {
-        int result = one * two;
+    public double multiplication(double one, double two) {
+        double result = one * two;
         return result;
     }
 
-    public int division(int one, int two) {
-        int result = one / two;
+    public double division(double one, double two) {
+        double result = one / two;
         return result;
     }
 }
